// packages/prisma-oppulence/generated/zod/inputTypeSchemas/PartnerMinOrderByAggregateInputSchema.ts
import type { Prisma } from '@prisma/client'

import { z } from 'zod'
import { SortOrderSchema } from './SortOrderSchema'

export const PartnerMinOrderByAggregateInputSchema: z.ZodType<Prisma.PartnerMinOrderByAggregateInput> = z
    .object({
        id: z.lazy(() => SortOrderSchema).optional(),
        name: z.lazy(() => SortOrderSchema).optional(),
        email: z.lazy(() => SortOrderSchema).optional(),
        image: z.lazy(() => SortOrderSchema).optional(),
        bio: z.lazy(() => SortOrderSchema).optional(),
        country: z.lazy(() => SortOrderSchema).optional(),
        status: z.lazy(() => SortOrderSchema).optional(),
        stripeConnectId: z.lazy(() => SortOrderSchema).optional(),
        payoutsEnabled: z.lazy(() => SortOrderSchema).optional(),
        showOnLeaderboard: z.lazy(() => SortOrderSchema).optional(),
        createdAt: z.lazy(() => SortOrderSchema).optional(),
        updatedAt: z.lazy(() => SortOrderSchema).optional()
    })
    .strict()

export default PartnerMinOrderByAggregateInputSchema